job-manager: track outstanding alloc/start requests for quiescense
after receiving an alloc response from the scheduler, the job-manager
emits an event, which triggers a `start` request to be sent to the exec
system.  The re-entrance into the reactor loop between the reception of
the alloc response and sending the start request means that the
job-manager has a chance to "pre-maturely" process the quiescent
response from the scheduler.  This ultimately leads to the simulator
receiving an erroneous 'quiescent' response from the job-manager.  A
similar problem exists for outstanding start requests.

To solve these problems, ensure that every alloc response has a
corresponding start response before sending a quiescent request.  Track
the number of outstanding requests in the simulator context of the job
manager, which is also the piece responsible for responding to the
quiescent request.